backend/pyproject.toml
[tool.poetry]
name = "museum"
version = "0.1.0"
description = ""
authors = ["Otto Zhang <otto@vianova.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = {version = "^0.30.1", extras = ["standard"]}
fastapi = "^0.110.3"
sqlalchemy = "1.4.52"
pydantic = "2.7.2"
pydantic-settings = "2.2.1"
pydantic_core = "2.18.3"
getconf = "^1.11.1"
psycopg2 = "^2.9.9"
postgis = "^1.0.4"
requests = "^2.32.3"
jwt = "^1.3.1"
passlib = "^1.7.4"
alembic = "^1.13.1"
bcrypt = "4.0.1"
pyjwt = "^2.8.0"
starlette-context = "^0.3.6"
casbin = "^1.36.2"
pytest-mock = "^3.14.0"
testing-postgresql = "^1.3.0"
pyyaml = "^6.0.1"
email-validator = "^2.2.0"
jinja2 = "^3.1.4"
httpx = "^0.27.0"
python-multipart = "^0.0.9"
openai = "^1.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.9.3"
black = "^24.4.2"
autoflake = "^2.3.1"
pytest = "^8.2.2"
pytest-cov = "5.0.0"
pytest-forked = "1.4.0"
pytest-golden = "0.2.2"
pytest-mock = "3.14.0"
pytest-rerunfailures = "14.0"
pytest-xdist = "3.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.direnv
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | museum/alembic/versions
  | explorer
  | node_modules
  | ansible
)
'''

[tool.flake8]
exclude = [
  "build",
  ".tox",
  ".direnv",
  ".eggs",
  ".venv",
  "alembic/versions",
  "explorer",
  "node_modules/",
  "ansible",
]
max-line-length = 88

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
src_paths = ["."]

[tool.pytest.ini_options]
markers = [
    "stream: marks tests relevant to streams",
]
